[kernels]: add common/mem_ops for shared memory/atomic helpers
Move get_llvm_ptr / atomic_add / pointer / global-load helpers into a
single kernels/common/mem_ops.py; kernels_common and utils re-export
them, so existing imports are unchanged. Add global_store and
buffer_atomic_add as single entry points, an __all__ documenting the
preferred vs legacy names, and an ir_value unwrap in global_store to
match atomic_add.

Route the hgemm_splitk semaphore through atomic_add, and every
raw_ptr_buffer_atomic_fadd site (MoE gemm/blockscale/mixed, conv3d and
conv3d_fp8, gfx1250 common) through buffer_atomic_add.

Codegen-neutral: final ISA is byte-identical before/after for every
touched kernel. gfx942 tests pass (hgemm_splitk, rmsnorm, moe gemm/
blockscale); gfx950 tests pass (moe gemm suite, conv3d, conv3d_fp8,
split-k flash).
