+++ kernels/moe/moe_gemm_2stage/gemm2.py
@@ -32,6 +32,7 @@ from flydsl._mlir import ir
 from flydsl._mlir.dialects import llvm, scf
 from flydsl.expr.typing import T
 from kernels.common.kernels_common import _if_then
+from kernels.common.mem_ops import buffer_atomic_add
 from kernels.mma.mfma_epilogues import c_shuffle_epilog, default_epilog
 from kernels.mma.mfma_preshuffle_pipeline import (
     buffer_copy_gmem16_dwordx4,
@@ -1164,13 +1165,7 @@ def compile_moe_gemm2(
                 e_vec = _e_vec
 
                 def atomic_add_f16x2(val_f16x2, byte_off_i32):
-                    rocdl.raw_ptr_buffer_atomic_fadd(
-                        val_f16x2,
-                        out_rsrc,
-                        byte_off_i32,
-                        zero_i32,
-                        zero_i32,
-                    )
+                    buffer_atomic_add(val_f16x2, out_rsrc, byte_off_i32, zero_i32, zero_i32)
 
                 sw_pf = None
                 tw_pf = None
@@ -1205,13 +1200,7 @@ def compile_moe_gemm2(
                     c4_i32 = fx.Int32(4)
 
                     def atomic_add_f32(val_f32, byte_off_i32):
-                        rocdl.raw_ptr_buffer_atomic_fadd(
-                            val_f32,
-                            out_rsrc,
-                            byte_off_i32,
-                            zero_i32,
-                            zero_i32,
-                        )
+                        buffer_atomic_add(val_f32, out_rsrc, byte_off_i32, zero_i32, zero_i32)
 
                     def _stage2_row_atomic(*, mi: int, ii: int, row_in_tile, row):
                         fused2 = buffer_ops.buffer_load(sorted_rsrc, row, vec_width=1, dtype=T.i32)
